Remove chunk manager
